refactor: refactored campaigns files into folders
