+++ api/src/resolvers/campaigns/equipment/equipment-campaign-cypher.ts
