+++ api/src/resolvers/campaigns/multiplication/multiplication-campaign-cypher.ts
