+++ api/src/resolvers/campaigns/sheep-seeking/sheep-seeking-resolvers.ts
@@ -1,6 +1,9 @@
-import { MakeServant, RemoveServant } from '../directory/make-remove-servants'
-import { permitAdmin } from '../permissions'
-import { Context } from '../utils/neo4j-types'
+import {
+  MakeServant,
+  RemoveServant,
+} from '../../directory/make-remove-servants'
+import { permitAdmin } from '../../permissions'
+import { Context } from '../../utils/neo4j-types'
 
 const sheepSeekingMutations = {
   MakeStreamSheepSeeker: async (object: never, args: never, context: Context) =>
