+++ api/src/resolvers/resolvers.ts
@@ -3,20 +3,14 @@ import serviceNoIncomeMutations from './no-income/service-resolvers'
 import serviceMutation from './services/service-resolvers'
 import { Member } from './utils/types'
 import treasuryMutations from './anagkazo/treasury-resolvers'
-import sheepSeekingMutations from './campaigns/sheep-seeking-resolvers'
 import directoryMutation from './directory/directory-resolvers'
 import {
   arrivalsMutation,
   arrivalsResolvers,
 } from './arrivals/arrivals-resolvers'
 import bankingMutation from './banking/banking-resolver'
-import multiplicationCampaignMutations from './campaigns/multiplication-campaign-resolvers'
 import campaignsResolvers from './campaigns/campaigns-resolver'
-import {
-  equipmentCampaignMutations,
-  equipmentCampaignResolvers,
-} from './campaigns/equipment-campaign-resolvers'
-import swollenSundayMutations from './campaigns/swollen-sunday-campaign-resolvers'
+import campaignMutations from './campaigns/campaign-mutations'
 
 const dotenv = require('dotenv')
 
@@ -53,20 +47,16 @@ const resolvers = {
   },
   Constituency: {
     ...campaignsResolvers.Constituency,
-    ...equipmentCampaignResolvers.Constituency,
   },
   Council: {
     ...campaignsResolvers.Council,
-    ...equipmentCampaignResolvers.Council,
   },
   Stream: {
     ...campaignsResolvers.Stream,
     ...arrivalsResolvers.Stream,
-    ...equipmentCampaignResolvers.Stream,
   },
   GatheringService: {
     ...campaignsResolvers.GatheringService,
-    ...equipmentCampaignResolvers.GatheringService,
   },
 
   Mutation: {
@@ -77,10 +67,7 @@ const resolvers = {
     ...bankingMutation,
     ...treasuryMutations,
     ...serviceNoIncomeMutations,
-    ...sheepSeekingMutations,
-    ...multiplicationCampaignMutations,
-    ...equipmentCampaignMutations,
-    ...swollenSundayMutations,
+    ...campaignMutations,
   },
 }
